# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kaiwa App API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/shadowing/{content_id}/in-progress": {
      "get": {
        "tags": ["Shadowing"],
        "summary": "Resume the latest in-progress Shadowing attempt",
        "operationId": "getInProgressShadowingAttempt",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Published shadowing content ID",
              "title": "Content Id"
            },
            "description": "Published shadowing content ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowingResumeResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/shadowing/{content_id}/record-segment": {
      "post": {
        "tags": ["Shadowing"],
        "summary": "Upload audio recording for a shadowing segment",
        "operationId": "recordShadowingSegment",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Published shadowing content ID",
              "title": "Content Id"
            },
            "description": "Published shadowing content ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recordShadowingSegment"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowingRecordSegmentResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shadowing/{content_id}/record-continuous": {
      "post": {
        "tags": ["Shadowing"],
        "summary": "Upload audio recording for continuous full-material shadowing",
        "operationId": "recordShadowingContinuous",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Published shadowing content ID",
              "title": "Content Id"
            },
            "description": "Published shadowing content ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recordShadowingContinuous"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowingRecordContinuousResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shadowing/{content_id}/submit": {
      "post": {
        "tags": ["Shadowing"],
        "summary": "Submit and finalize a shadowing attempt",
        "operationId": "submitShadowingAttempt",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Published shadowing content ID",
              "title": "Content Id"
            },
            "description": "Published shadowing content ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShadowingSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowingSubmitResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shadowing/attempts/{attempt_id}/review": {
      "get": {
        "tags": ["Shadowing"],
        "summary": "Review a Shadowing attempt with segment recordings",
        "operationId": "getShadowingAttemptReview",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Shadowing attempt ID",
              "title": "Attempt Id"
            },
            "description": "Shadowing attempt ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowingAttemptReviewResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shadowing/recordings/{recording_id}/playback": {
      "get": {
        "tags": ["Shadowing"],
        "summary": "Get presigned or public playback URL for a user recording",
        "operationId": "getShadowingRecordingPlayback",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Recording ID",
              "title": "Recording Id"
            },
            "description": "Recording ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowingRecordingPlaybackResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shadowing/recordings/{recording_id}": {
      "get": {
        "tags": ["Shadowing"],
        "summary": "Get presigned or public playback URL for a user recording",
        "operationId": "getShadowingRecording",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Recording ID",
              "title": "Recording Id"
            },
            "description": "Recording ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowingRecordingPlaybackResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttemptStatus": {
        "type": "string",
        "enum": ["in_progress", "completed"],
        "title": "AttemptStatus"
      },
...
      "Body_recordShadowingContinuous": {
        "properties": {
          "audio_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio File",
            "description": "User continuous audio recording file"
          },
          "attempt_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id",
            "description": "Optional attempt ID"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds",
            "description": "Optional continuous duration in seconds"
          }
        },
        "type": "object",
        "required": ["audio_file"],
        "title": "Body_recordShadowingContinuous"
      },
      "Body_recordShadowingSegment": {
        "properties": {
          "audio_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio File",
            "description": "User audio recording file"
          },
          "segment_id": {
            "type": "string",
            "title": "Segment Id",
            "description": "Segment identifier"
          },
          "attempt_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id",
            "description": "Optional attempt ID"
          }
        },
        "type": "object",
        "required": ["audio_file", "segment_id"],
        "title": "Body_recordShadowingSegment"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": ["error"],
        "title": "ErrorResponse",
        "description": "Stable envelope for expected and unexpected API errors."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShadowingContinuousRecordingSummary": {
        "properties": {
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": ["recording_id", "storage_key", "duration_seconds"],
        "title": "ShadowingContinuousRecordingSummary"
      },
...
      "ShadowingMode": {
        "type": "string",
        "enum": ["segmented", "continuous"],
        "title": "ShadowingMode"
      },
      "ShadowingRecordContinuousResponse": {
        "properties": {
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "attempt_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attempt Id"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "recording_id",
          "attempt_id",
          "storage_key",
          "duration_seconds",
          "created_at"
        ],
        "title": "ShadowingRecordContinuousResponse"
      },
      "ShadowingRecordSegmentResponse": {
        "properties": {
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "attempt_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attempt Id"
          },
          "segment_id": {
            "type": "string",
            "title": "Segment Id"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "recording_id",
          "attempt_id",
          "segment_id",
          "storage_key",
          "duration_seconds",
          "created_at"
        ],
        "title": "ShadowingRecordSegmentResponse"
      },
      "ShadowingRecordedSegmentSummary": {
        "properties": {
          "segment_id": {
            "type": "string",
            "title": "Segment Id"
          },
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "segment_id",
          "recording_id",
          "duration_seconds",
          "created_at"
        ],
        "title": "ShadowingRecordedSegmentSummary"
      },
      "ShadowingRecordingPlaybackResponse": {
        "properties": {
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "playback_url": {
            "type": "string",
            "title": "Playback Url"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "recording_id",
          "playback_url",
          "duration_seconds",
          "created_at"
        ],
        "title": "ShadowingRecordingPlaybackResponse"
      },
      "ShadowingResumeResponse": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attempt Id"
          },
          "content_id": {
            "type": "string",
            "format": "uuid",
            "title": "Content Id"
          },
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number"
          },
          "mode": {
            "$ref": "#/components/schemas/ShadowingMode",
            "default": "segmented"
          },
          "total_segments": {
            "type": "integer",
            "title": "Total Segments"
          },
          "recorded_segments": {
            "items": {
              "$ref": "#/components/schemas/ShadowingRecordedSegmentSummary"
            },
            "type": "array",
            "title": "Recorded Segments"
          },
          "continuous_recording": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShadowingContinuousRecordingSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "total_attempts": {
            "type": "integer",
            "title": "Total Attempts",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "content_id",
          "attempt_number",
          "total_segments"
        ],
        "title": "ShadowingResumeResponse"
      },
...
      "ShadowingSubmitRequest": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attempt Id"
          },
          "replay_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Replay Count",
            "default": 0
          }
        },
        "type": "object",
        "required": ["attempt_id"],
        "title": "ShadowingSubmitRequest"
      },
      "ShadowingSubmitResponse": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attempt Id"
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "xp_earned": {
            "type": "integer",
            "title": "Xp Earned"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "shadowing"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "B\u1ea1n \u0111\u00e3 ho\u00e0n th\u00e0nh b\u00e0i luy\u1ec7n."
          },
          "user_progress": {
            "$ref": "#/components/schemas/ShadowingUserProgressSummary"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          },
          "ai_feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShadowingAiFeedback"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "status",
          "score",
          "xp_earned",
          "difficulty",
          "user_progress",
          "completed_at"
        ],
        "title": "ShadowingSubmitResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}